Restructure and product document spec.
