+++ eo3/metadata/validate.py
